refactor: remove plural type defs
